server/requirements.txt
blinker==1.9.0
click==8.1.8
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
tzdata==2025.2
Werkzeug==3.1.3
